pyproject.toml
[tool.poetry]
name = "comsol"
version = "0.1.0"
description = ""
authors = ["banananafish <darkbananafish@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
MPh = "^1.2.3"
numpy = "^1.26.4"
pandas = "^2.2.1"
pyinstaller = "^6.5.0"
typing-extensions = "^4.10.0"
loguru = "^0.7.2"
click = "^8.1.7"
pyyaml = "^6.0.1"
rich = "^13.7.1"
pygad = "^3.3.1"
scikit-learn = "^1.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
tox = "^4.14.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
comsol = "comsol.cmdline:main"
